multifloats: MKL/OpenBLAS-style ILP64 (_64_) entry points alongside LP64
Add 64-bit-integer ILP64 entry points (name_64_) for all 75 public
multifloats routines (m = real double-double, w = complex DD), alongside
the existing LP64 name_ entries in the same archive, via the facade model
already used for kind10 and kind16: one ptrdiff_t-based core per routine
plus two thin facades (name_/name_64_) that deref-and-tail-call it. The
facade is the sole Fortran by-ref boundary, so it adds no measurable cost.

As a C++ target each EPBLAS_FACADE_* invocation is wrapped in extern "C" {}
to keep C ABI on the two facade symbols and the core. The double-double
structs are exact-sized (16 B / 32 B), so output memcmp is exact.

Verification: 75/75 LP64/ILP64 symbol pairs (0 unpaired); A/B twin oracle
bit-identical at OMP=1 and OMP=4; 75/75 fuzz pass; before/after dual-link
perf 313 cells, 0 regressions (medians ~1.00).

+++ src/epblas-parallel/multifloats/mf_omp.h
@@ -135,14 +135,15 @@ static inline void band_row_window(std::ptrdiff_t t, bool upper, const std::ptrd
 
 /* even_slice(): the flat EQUAL-COUNT partition of [0,n) used by every threaded L1
  * / reduction sweep (copy/scal/axpy/rot/asum/dot/nrm2/argmax). Thread tid of nth
- * gets [lo,hi) = [n*tid/nth, n*(tid+1)/nth); the long-long product avoids std::ptrdiff_t
- * overflow before the divide. Byte-identical math previously re-derived in ~20
- * files — centralized here so the slice bounds (and their overflow guard) live
- * once. */
+ * gets [lo,hi) = [n*tid/nth, n*(tid+1)/nth); the product is formed in __int128
+ * so the multiply cannot overflow std::ptrdiff_t before the divide (ILP64-safe —
+ * the same guard as common/blas_omp.h's blas_part_bound). Byte-identical math
+ * previously re-derived in ~20 files — centralized here so the slice bounds (and
+ * their overflow guard) live once. */
 static inline void even_slice(std::ptrdiff_t n, std::ptrdiff_t tid, std::ptrdiff_t nth, std::ptrdiff_t &lo, std::ptrdiff_t &hi)
 {
-    lo = (std::ptrdiff_t)((long long)n * tid / nth);
-    hi = (std::ptrdiff_t)((long long)n * (tid + 1) / nth);
+    lo = (std::ptrdiff_t)((__int128)n * tid / nth);
+    hi = (std::ptrdiff_t)((__int128)n * (tid + 1) / nth);
 }
 
 }  // namespace mf_omp
